Fix store persistence misconfiguration
