--- packages/data-stores/src/help-center/persist.ts
@@ -1,3 +1,0 @@
-import persistenceConfigFactory from '../persistence-config-factory';
-
-export default persistenceConfigFactory( 'WP_HELP_CENTER' );
