+++ packages/data-stores/src/index.ts
@@ -6,7 +6,7 @@ import * as HelpCenter from './help-center';
 import * as I18n from './i18n';
 import * as Launch from './launch';
 import * as Onboard from './onboard';
-import persistenceConfigFactory from './persistence-config-factory';
+import oneWeekPersistenceConfig from './one-week-persistence-config';
 import * as Plans from './plans';
 import * as ProductsList from './products-list';
 import * as Reader from './reader';
@@ -46,7 +46,7 @@ export {
 	WPCOMFeatures,
 	Reader,
 	Onboard,
-	persistenceConfigFactory,
+	oneWeekPersistenceConfig,
 	ProductsList,
 	AutomatedTransferEligibility,
 	StepperInternal,
