--- packages/data-stores/src/launch/persist.ts
@@ -1,3 +1,0 @@
-import persistenceConfigFactory from '../persistence-config-factory';
-
-export default persistenceConfigFactory( 'WP_LAUNCH' );
