+++ packages/data-stores/src/onboard/index.ts
@@ -1,8 +1,8 @@
 import { plugins, registerStore, use } from '@wordpress/data';
 import { controls } from '@wordpress/data-controls';
+import persistOptions from '../one-week-persistence-config';
 import * as actions from './actions';
 import { STORE_KEY } from './constants';
-import persistOptions from './persist';
 import reducer, { State } from './reducer';
 import * as selectors from './selectors';
 import type { SelectFromMap, DispatchFromMap } from '../mapped-types';
@@ -11,11 +11,15 @@ export type { State };
 
 export { SiteGoal, SiteIntent } from './constants';
 export * as utils from './utils';
+let isRegistered = false;
 
 /**
  * Onboard store depends on site-store. You should register the site before using this store.
  */
 export function register(): typeof STORE_KEY {
+	if ( isRegistered ) {
+		return STORE_KEY;
+	}
 	use( plugins.persistence, persistOptions );
 
 	registerStore< State >( STORE_KEY, {
@@ -48,7 +52,7 @@ export function register(): typeof STORE_KEY {
 			'storeType',
 		],
 	} );
-
+	isRegistered = true;
 	return STORE_KEY;
 }
