--- packages/data-stores/src/onboard/persist.ts
@@ -1,3 +1,0 @@
-import persistenceConfigFactory from '../persistence-config-factory';
-
-export default persistenceConfigFactory( 'WP_ONBOARD' );
